requirements.txt
google-cloud-pubsub==2.19.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
